Simply daemonset creation
The daemonsets were created as workers and masters separately. While
this is fine. This refactor attempts to simplify the code by issuing one
daemonset for every node that matches.

There is already a toleration for the master nodes, so every node in the
cluster will match and have a daemonset scheduled.

On a subsequent iteration, we can add an interface to filter the nodes
per label. But that will be done in a subsequent PR.