Tie AIDE config to FileIntegrity Object
This ties the AIDE ConfigMap to the FileIntegrity object by giving them
the same name. Effectively, this ends up making the configurations
unique per file-integrity object, as opposed to relying on a single
ConfigMap.

This is the first step on enabling us to use multiple FileIntegrities
per deployment (which would ideally serve different node sets).

Given that the name of the config file matches the name of the
FileIntegrity, we can infer the DaemonSet name from it. Which are also
now unique per CR instance.