Add annotation to force datbase re-init
This creates an annotation called "file-integrity.openshift.io/re-init"
that if it's present in the FileIntegrity object, it'll spawn the
re-init daemonset, thus reinitializing the database.

This is useful in cases where we have done expected changes in the
filesystem and we want to get rid of the warnings the operator is doing.