FI ctrl: Expose the period between scans as a new CRD attribute gracePeriod
Extends the FileIntegrityConfig structure with a new member GracePeriod. If
not set, its value defaults to 10 seconds, otherwise the configured
value is used for both the pause between individual scans of the aide
integrity scanner and the interval in which the logcollector checks for
aide results.

Because the grace Period customizes the aide script, the script is no
longer a global shared singleton, but a per-FI configmap.

When the FI object's gracePeriod is changed, the per-script configmap is
changed, labeled with a label key that signifies that the configmap had
been updated and the CM controller takes care of restarting the DS to
pick up the updated configmap.

Jira: CMP-436