e2e: Fix label mismatch
The code was using the same label to match nodes and to assign a MC to a
pool. That caused the MC to never be applied, because an mc must use the
machineconfiguration.openshift.io/role role that ties it to a pool, but
nodes are matches with node-role.kubernetes.io/worker.