util: A race might delete a reinit DS before it starts
The FI controller would on reinit create a reinit-DS and annotate the
config script. This annotation would be picked up by the CM controller
which checks if the reinit DS was done by checking if
desiredNumberScheduled is equal to NumberAvailable.

But if the CM woke up before the DS was scheduled, both
desiredNumberScheduled and NumberAvailable were zero, so the test
thought the reinit DS already ran while in fact it wasn't even
scheduled.