CM: Fix deleteDaemonSetPods()
deleteDaemonSetPods() was filtering the pods in the namespace by
ds.GetLabels(), but the DS we create has no labels, so in effect the
resulting labelSelector matched all pods in the namespace and the
function would just delete all pods in the namespace, including the
operator.

This patch uses the "app: $dsName" label instead.