Make default grace period to be 15 min
The AIDE integrity checking is a very intensive process, with our
current default of 10s, the daemon would constantly take ~2.2G of memory
(for each instance... meaning, in each node).

This is excessive and will certaintly raise alarms with folks running
this in production.

I propose to set our default grace period to be 900 (15 min), which
will result in the daemonsets taking 60MB when idle, and having usage
spikes every 15 mins.