Bug 1861303: Migrate holdoff logic to daemonset
Having separate logic in the node controller to holdoff the integrity
checks, this patch migrates that functionality to the daemon itself.

This instead adds a watcher for the file integrity instance as part of
the daemon (having its own loop), and now the holdoff functionality
solely relies on the fileintegrity instance having the appropriate
annotations. So no need to check files anymore.