deamon: Safer utility function to read the return code
The way we read the return code of AIDE had several issues. It wouldn't
work if the command exited uncleanly (like, with a signal) and even
worse, the default exitStatus value was 0, so in case any of the type
casts didn't work, we would just have assumed zero.

This patch adds a utility function common.GetAideExitCode() which is a
bit safer in the sense that it is able to detect an unclean exit and in
that case returns a fallback error code. In this patch, the fallback
error code is a made-up magic constant, this will be fixed in a
follow-up patch.