Reload FI when the configMap the FI points to changes
The FI controller used to only watch for FI object changes. That meant
that if the configMap pointed to by spec.config changed, the operator
would not re-render the configuration the DS uses until the FI object
was 'touched'.

This patch uses a custom mapper to reconcile FI objects when CMs that
are used by a FI changes.