logcollector: Only read log once in a buffered way
Instead of reading the whole file and executing operations on top of
that loaded file. This loads the reader interface on to a buffered
reader and passes that around; thus only reading the file when it's
really needed (for compression or writing).

This also removes the offset tracking, which wasn't being used since
AIDE truncates the log file before writing it.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>