Use an absolute git reference for the upstream repository
Before, we assumed `origin` always pointed to the upstream repository
for the operator (openshift/file-integrity-operator). This might not be
the case if someone sets `origin` to their own fork and uses `upstream`
to point to openshift/file-integrity-operator.

Either way, we can generalize the git push commands to use an
absolute-like reference for the remote.

Fixes #246