Update golang version in Dockerfile for builds
We have a separate Dockerfile used for building container images during
the release process, which is separate from the Dockerfile we use for CI
(Dockerfile.ci).

We recently updated the version of golang we use to 1.19 by updating
go.mod and the Dockerfile.ci file. But, we didn't update the golang
version for the builds we use during the upstream release process.

I caught this issue as I prepared the 1.2.0 release and noticed failures
building the new container images for that version.

This commit updates the golang version in build/Dockerfile to be
consistent with the version use in the rest of the project.