Use remote detection in release process
Some contributors use different naming conventions and protocols for
their upstream and downstream branchs. Instead of assuming the same
setup for everyone, let's use a script to find which remote name to use
based on the organization and repository, and then call that script from
the release targets.

This makes the release process a little easier for contributors,
regardless of how they have their remotes setup.