Update aide config to support aide 0.18.4
Newer versions of fedora use aide 0.18.0, instead of 0.16.0. Previously,
we were using a config setting called `verbose` which isn't supported in
newer versions of aide.

This commit updates the aide config to work with newer versions of aide.
We shouldn't have to worry about supporting two different versions
of aide since the aide version is baked into the images shipped with the
operator.

Fixes #396