Pin aide version to 0.16.0
FIO makes some assumption about aide configuration. Newer version of
aide have backwards incompatible changes with older configurations that
FIO is using, causing issues with CI. We will need to handle the
configuration migration in the operator to allow for smooth migration.

This commit pins the version of aide we're using to 0.16.0 for now until
we can handle the configuration migration and use a newer version of
aide.

Fixes #396