AIDE migration
This PR adds ability to run aide 0.18 config checks and migration.

The aide0.18 will be complied as a CGO library and used by FIO, we
will try to see if we can perform the migration if a user defined config is being
detected. We will issue warning messages in the log, and have a failed annotation key
in the FIO instance if we are not able to pre migrate the config.