Only set node selectors for OpenShift deployments
By default, the File Integrity Operator would set node selectors and
taints that would only allow it to schedule on `master` nodes. While
this is fine for OpenShift environments, some additional environments we
test on, like Red Hat OpenShift on AWS (ROSA), don't provide `master`
nodes at all.

To accomodate this in our testing and tooling, this commit removes the
node selector and taint from the default bundle, and moves it into an
OpenShift-specific bundle where it will still get used for OpenShift
deploy paths and bundles.

This will be more useful in a future patch that implements support for
running File Integrity Operator end-to-end testing on ROSA HCP, which
only provides `worker` nodes and not `master` nodes.