CMP-2624: Implement support for running e2e tests on ROSA
This commit adds some test plumbing so that the e2e framework knows when
it's running on ROSA. This is necessary so that the framework bypasses
certain features that aren't supported on that platform, like Machine
Configs and Machine Config Pools.

By implementing this, we're able to integrate ROSA CI into the gating
process to ensure the operator works on managed OpenShift offerings.