Skip test that require machine configs on ROSA HCP
We recently added support for running tests on ROSA HCP, which doesn't
support some objects like Machine Configs:

  https://github.com/openshift/release/pull/53502

Because of this, the TestFileIntegrityAcceptsExpectedChange test is
failing because it expects to create a Machine Config as a dependency
for the test.

This commit updates the test to check if we're running on ROSA HCP, and
if so, skip the test.