Add a Makefile target to verify bundles
We don't have any CI that checks to make sure the bundle stays
up-to-date with PRs that may be changing the bundle. This usually
results in dirty local builds because running `make bundle` will
generate new bundles from what is kept in source based on other parts
othe project, or the operator-sdk version.

This commit adds in a target to run the `bundle` and then verify there
haven't been any changes. We can wire this up to CI to make it easier to
catch bundle changes sooner.