CMP-3362: Update bundle images for Red Hat registries
This commit updates the bundle build scripts so that it references
images in registry.redhat.io, so that the bundle image will work once
it's released. The images won't resolve, and we will use an
ImageContentSourcePolicy to replace the image requests to
registry.redhat.io with the quay.io references maintained by Konflux.

The important part here is that we're using the same SHA as what's
proposed in quay.io, so that when the image is mirrored the reference
will resolve.